refactor task creation flow
