+++ styles.css
@@ -1,3 +1,5 @@
+@import url('https://fonts.googleapis.com/css2?family=Open+Sans:wght@700&display=swap');
+
 @font-face {
   font-family: 'Lato';
   src: url('Fonts/Lato-Regular.ttf') format('truetype');
@@ -201,10 +203,6 @@ li:hover { transform: scale(1.02); }
   transition: transform 0.3s ease;
 }
 
-.menu-carousel span {
-  margin-top: 10px;
-}
-
 .menu-grid {
   display: grid;
   gap: 40px;
@@ -395,8 +393,8 @@ li:hover { transform: scale(1.02); }
 }
 
 .task-form {
-  background: var(--bg-color);
-  color: var(--text-color);
+  background: linear-gradient(135deg, #40E0D0, #1E90FF);
+  color: #fff;
   padding: 20px;
   border-radius: 8px;
   width: calc(100% - 60px);
@@ -404,17 +402,55 @@ li:hover { transform: scale(1.02); }
   display: flex;
   flex-direction: column;
   gap: 10px;
+  font-family: 'Open Sans', sans-serif;
+  font-weight: 700;
 }
 
-.task-form input,
-.task-form textarea {
+.task-form input {
   padding: 8px;
-  border: 1px solid #ccc;
+  border: none;
   border-radius: 4px;
+  background: rgba(255,255,255,0.2);
+  color: #fff;
+}
+
+.task-form select {
+  padding: 8px;
+  border: none;
+  border-radius: 4px;
+  background: #40E0D0;
+  color: #fff;
 }
 
 .task-form button {
   width: 100%;
+}
+
+#to-step-2, #to-step-3, #save-task {
+  background: linear-gradient(135deg, #00ff7f, #32cd32);
+  color: #fff;
+}
+
+#cancel-task {
+  background: linear-gradient(135deg, #ff6347, #ff0000);
+  color: #fff;
+}
+
+.task-form .task-step > input,
+.task-form .task-step > select,
+.task-form .task-step > textarea,
+.task-form .task-step > .time-inputs {
+  width: calc(100% - 40px);
+  margin: 0 20px;
+}
+
+.task-form .time-inputs {
+  display: flex;
+  gap: 10px;
+}
+
+.task-form .time-inputs input {
+  flex: 1;
 }
 
 .modal-buttons {
